subprocess.Popen
And there there have it. The Python script will run the PowerShell script from an event that happens in the Python GUI.

Ya I learned a new way to use an old word. It's helpful as well in my thinking as everything that happens on a computer is because of an event. I find this helpful in thinking about how to program.

Next step understanding how I am interacting with the event and how to pass stdin from Python to PowerShell.