I new I was done when I changed the them for the 8th time...
learning, and I did that so whatev's

The new term today is tags. I can tag an input button with key='name' and then use that value to update a value. Or for something else, I presume, but that is what I used it for today.

Lastly, I discovered that this module has themes and played with those for a while.

It is neat to come into this from my PowerShell background. I find I am connecting processes in Python to ones I know in PowerShell. Also, seeing the same process in a new light helps me better understand the process.